+++ seantis/people/__init__.py
@@ -1,2 +1,8 @@
+import logging
+log = logging.getLogger('seantis.people')
+
 from zope.i18nmessageid import MessageFactory
 _ = MessageFactory('seantis.people')
+
+# duplicated in profiles/default/toolset/toolset.xml
+catalog_id = 'seantis_people_catalog'