+++ seantis/people/setuphandler.py
@@ -1,3 +1,4 @@
+from seantis.people import catalog_id
 from seantis.plonetools import setuphandlers
 
 indexes = [
@@ -8,12 +9,18 @@ indexes = [
 
 
 def add_catalog_indexes(context, logger=None):
-    setuphandlers.add_catalog_indexes(
-        'seantis.people', indexes, context, logger
-    )
+    # these indexes are not dynamic and need to be available in both catalogs
+    # -> only the dynamic indexes created by the supermodel are only bound
+    # to the people's catalog
+    for catalog in ('portal_catalog', catalog_id):
+        setuphandlers.add_catalog_indexes(
+            'seantis.people', indexes, context, logger, catalog
+        )
 
 
 def import_indexes(context):
-    setuphandlers.import_indexes(
-        'seantis.people', indexes, context
-    )
+    # see above..
+    for catalog in ('portal_catalog', catalog_id):
+        setuphandlers.import_indexes(
+            'seantis.people', indexes, context, catalog
+        )