+++ seantis/people/tests/test_order.py
@@ -5,6 +5,7 @@ from seantis.people.supermodel import (
 )
 from seantis.plonetools import tools
 
+from seantis.people import catalog_id
 from seantis.people.interfaces import IPerson
 from seantis.people import tests
 
@@ -12,7 +13,10 @@ from seantis.people import tests
 class TestOrder(tests.IntegrationTestCase):
 
     def test_order(self):
-        new_type = self.new_temporary_type(behaviors=[IPerson.__identifier__])
+        new_type = self.new_temporary_type(
+            behaviors=[IPerson.__identifier__],
+            klass='seantis.people.types.base.PersonBase'
+        )
         set_order(new_type.lookupSchema(), ['bar', 'foo'])
 
         with self.user('admin'):
@@ -24,8 +28,8 @@ class TestOrder(tests.IntegrationTestCase):
                 bar='hammertime!'
             )
 
-        brain = tools.get_brain_by_object(obj)
-        catalog = api.portal.get_tool('portal_catalog')
+        brain = tools.get_brain_by_object(obj, catalog_id)
+        catalog = api.portal.get_tool(catalog_id)
 
         self.assertEqual(
             catalog.getIndexDataForRID(brain.getRID())['sortable_title'],